+++ tests/phpunit/Thelia/Tests/ImportExport/Export/ProductPricesExportTest.php
@@ -67,9 +67,8 @@ class ProductPricesExportTest extends \PHPUnit_Framework_TestCase
             $this->assertNotNull($currency);
 
             $price = $pse->getPricesByCurrency($currency);
-            // The substr is a patch for php 5.4 float round
-            $this->assertEquals(substr($price->getPrice(), 0, strlen($row["price"])), $row["price"]);
-            $this->assertEquals(substr($price->getPromoPrice(), 0, strlen($row["promo_price"])), $row["promo_price"]);
+            $this->assertEquals($price->getPrice(), $row["price"]);
+            $this->assertEquals($price->getPromoPrice(), $row["promo_price"]);
             $this->assertEquals($pse->getProduct()->getTitle(), $row["title"]);
 
             $attributeCombinations = $pse->getAttributeCombinations();